requirements.txt
Flask==3.0.0
SQLAlchemy==2.0.23
psycopg2-binary==2.9.9
websockets==12.0
fastapi==0.104.1
uvicorn==0.24.0
yfinance==0.2.31
requests==2.31.0
werkzeug==3.0.1
python-dotenv==1.0.0
iqoptionapi
pandas==2.1.3
numpy==1.26.2
flask-cors
flask-sock
redis
websocket-client
pillow
websocket
